refactor: 💡 lesson domain 커스텀 에러 패키지 분리
