+++ src/main/java/gwasuwonshot/tutice/lesson/exception/invalid/InvalidLessonCodeException.java
@@ -1,4 +1,4 @@
-package gwasuwonshot.tutice.lesson.exception;
+package gwasuwonshot.tutice.lesson.exception.invalid;
 
 import gwasuwonshot.tutice.common.exception.BasicException;
 import gwasuwonshot.tutice.common.exception.ErrorStatus;
