+++ src/main/java/gwasuwonshot/tutice/lesson/exception/invalid/InvalidTimeException.java
@@ -1,12 +1,13 @@
-package gwasuwonshot.tutice.lesson.exception;
+package gwasuwonshot.tutice.lesson.exception.invalid;
 
 import gwasuwonshot.tutice.common.exception.BasicException;
 import gwasuwonshot.tutice.common.exception.ErrorStatus;
 import lombok.Getter;
 
+
 @Getter
-public class InvalidDateException extends BasicException {
-    public InvalidDateException(ErrorStatus errorStatus, String message) {
+public class InvalidTimeException extends BasicException {
+    public InvalidTimeException(ErrorStatus errorStatus, String message) {
         super(errorStatus, message);
     }
